pyproject.toml
[project]
name = "demo-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.123.5,<0.124.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "pydantic[email]",
    "python-multipart",
    "PyJWT (>=2.0.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
